fix(security): SecretStr broker keys + Decimal financial + mypy core.models
Sprint 2 of post-audit cleanup. Three coordinated fixes:

1. Broker API keys (Alpaca, Binance), timescale_password, alert_smtp_password,
   twilio_sid/token, and db_password now use SecretStr instead of plain str
   (closes #71). All call sites updated to use .get_secret_value().

2. Removed mypy ignore_errors for core.models.* (closes #69). Zero errors
   surfaced — models were already well-typed. No new type: ignore comments.

3. Replaced float() with Decimal for financial values in core/models
   (TradeRecord.r_multiple, Signal.risk_reward) and S05 risk manager
   computation paths (closes #66). float() kept only at serialization
   boundaries (RuleResult kwargs, CircuitBreakerSnapshot.daily_loss_pct).

Files touched (14):
  core/config.py, core/models/order.py, core/models/signal.py,
  pyproject.toml, services/s01_data_ingestion/service.py,
  services/s01_data_ingestion/connectors/alpaca_historical.py,
  services/s05_risk_manager/circuit_breaker.py,
  services/s05_risk_manager/exposure_monitor.py,
  services/s05_risk_manager/position_rules.py,
  services/s06_execution/service.py, services/s10_monitor/alert_engine.py,
  tests/unit/test_config.py, docs/claude_memory/{SESSIONS,CONTEXT}.md

Files deferred:
  - S01 connectors (FRED, ECB, BoJ, EDGAR, SimFin): MacroPoint.value and
    FundamentalPoint.value model fields are typed float — requires model migration
  - S07 quant_analytics: float() for statistical computations (not financial)
  - S09 trade_analyzer: dedicated metrics sprint
  - S10 command_api: JSON serialization (not financial)

Quality gates:
  - ruff: clean
  - mypy --strict: 319 files, 0 errors
  - pytest: 1228 tests passing, zero regressions

Closes #66, closes #69, closes #71

+++ core/models/order.py
@@ -207,12 +207,12 @@ class TradeRecord(BaseModel):
         return self.net_pnl > 0
 
     @property
-    def r_multiple(self) -> float | None:
+    def r_multiple(self) -> Decimal | None:
         """Return PnL as multiple of initial risk (R)."""
         risk = abs(self.entry_price - self.exit_price)
         if risk == 0 or self.size == 0:
             return None
-        return float(self.net_pnl / (risk * self.size))
+        return self.net_pnl / (risk * self.size)
 
 
 class NullOrder(BaseModel):
